Calculates the change

calc_change calculates the level change of a column in a dataframe over a specified amount of time, calc_change_per calculates the percentage change of a column in a dataframe over a specified amount of time.

Usage

calc_change(column, len, dt)

calc_change_per(column, len, dt)

Arguments

column

The column to use for calculating the change.

len

The number of lags to use.

dt

The date column to order the column values by.

Value

A vector containing the change values.

Examples

library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

lfs_state %>%
group_by(date) %>%
summarise(emp = sum(emp), .groups = 'drop') %>%
mutate(chng = calc_change(emp, 1, date))
#> # A tibble: 152 × 3
#>    date         emp   chng
#>    <date>     <dbl>  <dbl>
#>  1 1986-08-01 6919.  NA   
#>  2 1986-11-01 7034. 116.  
#>  3 1987-02-01 7022. -11.9 
#>  4 1987-05-01 7104.  81.3 
#>  5 1987-08-01 7092. -11.3 
#>  6 1987-11-01 7182.  89.9 
#>  7 1988-02-01 7247.  65.1 
#>  8 1988-05-01 7355. 108.  
#>  9 1988-08-01 7353.  -1.55
#> 10 1988-11-01 7509. 155.  
#> # ℹ 142 more rows