Comparing models

Comparing models

Usage

compare_models(models, model_names = NULL)

Arguments

models

A statistical model object or a list of statistical model objects. Lists of models can be specified as models = list(model.1, model.2, ...). Different object types can also be mixed.

model_names

A character vector of labels for the model names. By default, the models are named "mod_1", "mod_2", etc. Specifying model_names = c("Model name 1", "Model name 2") etc. overrides the default behavior.

Value

A dataframe of compared models for easy model comparison

Examples

library(plm)
#> 
#> Attaching package: ‘plm’
#> The following objects are masked from ‘package:dplyr’:
#> 
#>     between, lag, lead
library(dplyr)

data(worldprod, package = 'jsar')
countries <- c("Australia", "singapore", "china", "uk")

# comparison with same model (lm) for each country
lm_models <- worldprod |>
  filter(country_name %in% countries) |>
  group_by(country_name) |>
  group_map(~ lm(log_gdp ~ log_capital + log_labour , data = .x))

compare_models(lm_models, model_names = paste0('lm_',countries))
#> # A tibble: 7 × 5
#>   term        lm_Australia  lm_singapore  lm_china      lm_uk        
#>   <chr>       <chr>         <chr>         <chr>         <chr>        
#> 1 (Intercept) -0.518(0.177) -9.019(3.665) 0.754(0.127)  -0.462(1.091)
#> 2 log_capital 1.36(0.174)   0.379(0.376)  0.346(0.097)* 0.51(0.027)**
#> 3 log_labour  -1.297(0.412) 1.828(0.921)  1.139(0.445)  0.862(0.351) 
#> 4 ####        ####          ####          ####          ####         
#> 5 R$^2$       0.996         0.94          0.996         0.986        
#> 6 Adj. R$^2$  0.995         0.935         0.995         0.985        
#> 7 Num. obs.   28            28            28            28           

# or with piping
worldprod |>
  filter(country_name %in% countries) |>
  group_by(country_name) |>
  group_map(~ lm(log_gdp ~ log_capital + log_labour , data = .x)) |>
  compare_models( model_names = paste0('lm_',countries))
#> # A tibble: 7 × 5
#>   term        lm_Australia  lm_singapore  lm_china      lm_uk        
#>   <chr>       <chr>         <chr>         <chr>         <chr>        
#> 1 (Intercept) -0.518(0.177) -9.019(3.665) 0.754(0.127)  -0.462(1.091)
#> 2 log_capital 1.36(0.174)   0.379(0.376)  0.346(0.097)* 0.51(0.027)**
#> 3 log_labour  -1.297(0.412) 1.828(0.921)  1.139(0.445)  0.862(0.351) 
#> 4 ####        ####          ####          ####          ####         
#> 5 R$^2$       0.996         0.94          0.996         0.986        
#> 6 Adj. R$^2$  0.995         0.935         0.995         0.985        
#> 7 Num. obs.   28            28            28            28           

lm_model <- lm(log_gdp ~ log_capital + log_labour, data = worldprod)
plm_model <- plm::plm(log_gdp ~ log_capital + log_labour ,
                      model = 'within', index = c('country_name', 'year'),
                      data = worldprod)

compare_models(list(lm_model, plm_model),model_names = c('Pool', 'Panel'))
#> # A tibble: 7 × 3
#>   term        Pool            Panel         
#>   <chr>       <chr>           <chr>         
#> 1 (Intercept) -0.577(0.016)** -             
#> 2 log_capital 0.862(0.005)*** 0.586(0.011)**
#> 3 log_labour  0.137(0.007)*** 0.291(0.028)**
#> 4 ####        ####            ####          
#> 5 R$^2$       0.963           0.89          
#> 6 Adj. R$^2$  0.963           0.886         
#> 7 Num. obs.   2296            2296