Comparing models

Comparing models

Usage

compare_models(models, model_names = NULL)

Arguments

models

an object containing the results returned by a model fitting function. Supported models are 'lm' & 'sfa', put multiple models in list

model_names

vector of model names as string

Value

dataframe of compared models and list for individual models

Examples

mod1 <- lm(Volume ~ Girth + Height, data = datasets::trees)
mod2 <- lm(Volume ~ Girth , data = datasets::trees)
# Comparing models
compare_models(list(mod1, mod2))$compare
#>          term               mod_1               mod_2
#> 1 (Intercept) -57.9877(8.6382)*** -36.9435(3.3651)***
#> 2       Girth   4.7082(0.2643)***   5.0659(0.2474)***
#> 3      Height    0.3393(0.1302)**                    
#> 4         rsq               0.948              0.9353
#> 5     adj_rsq              0.9442              0.9331
#> 6      f_stat         254.9723***         419.3603***
compare_models(mod2)
#> $compare
#>          term               mod_1
#> 1 (Intercept) -36.9435(3.3651)***
#> 2       Girth   5.0659(0.2474)***
#> 3         rsq              0.9353
#> 4     adj_rsq              0.9331
#> 5      f_stat         419.3603***
#> 
#> $models
#> $models[[1]]
#> # A tibble: 5 × 7
#>   term        estimate std.error statistic p.value sig   model         
#>   <chr>          <dbl>     <dbl>     <dbl>   <dbl> <chr> <chr>         
#> 1 (Intercept)  -36.9       3.37      -11.0       0 "***" Volume ~ Girth
#> 2 Girth          5.07      0.247      20.5       0 "***" Volume ~ Girth
#> 3 rsq            0.935    NA          NA        NA ""    Volume ~ Girth
#> 4 adj_rsq        0.933    NA          NA        NA ""    Volume ~ Girth
#> 5 f_stat       419.       NA          NA         0 "***" Volume ~ Girth
#> 
#>